Add `enum` and `const` to statements; tweak block parsing
Blocks will now parse a leading "description" comment (a triple-hyphen
comment) and assign that an AST value, as well as "directives", which
allow the compiler to read information about the source, such as whether
the file is a module and other information.

Enumerations and constants will be given purpose in later commits.