Add ability to rename variables during destructuring
This gives the ability to destructure values into a name that doesn't
represent how they were named in the previous table. This is useful in
cases where, for instance, a `const` value is assigned with the same
name as a value being imported from a module: the value imported from a
module can be given a different name, meaning the `const` value isn't
overridden.

Files changed:
- fusion/core/compilers/source.lua
- fusion/core/parser.lua
- spec/lexer_spec.lua
- spec/in/assignment.fuse
- spec/out/source/assignment.lua